create the details page and the _layout with stack. change the old _layout file to index
