+++ henriqueApp/app/details.tsx
@@ -1,0 +1,17 @@
+import { View, Text, StyleSheet } from 'react-native';
+
+export default function DetailsScreen() {
+  return (
+    <View style={styles.container}>
+      <Text>Details</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+});
